[Win] Moved EP error codes implementation into common directory.
This patch moved the error codes implementation from EP to the common
directory, and fixed up the build system and namespace accordingly.
We will write the error codes for SP and Face based on this.

This is the first step to move common codes into realsense/common
directory, then we will try to extract some common implementation or
usage (like SenseManager, etc.) from each extension and write them
here.